Moved Login and Header Components to Components Folder
