+++ weather-app/src/app/modules/app/home/home.component.ts
@@ -1,7 +1,7 @@
 import { Component } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { RouterModule } from '@angular/router';
-import { HeaderComponent } from './header/header.component';
+import { HeaderComponent } from '../components/header/header.component';
 
 @Component({
   selector: 'app-home',
