+++ weather-app/src/app/modules/app/home/home.route.ts
@@ -1,6 +1,6 @@
 import { Routes } from "@angular/router";
 import { HomeComponent } from "./home.component";
-import { LoginComponent } from "./login/login.component";
+import { LoginComponent } from "../components/login/login.component";
 
 export const HOME_ROUTE: Routes = [
     { path: '', component: HomeComponent},
